Fix native build by relaxing pdnsd compiler flags
- Reverted `pdnsd` compilation flags in `native/CMakeLists.txt` to remove `-Werror` and `-Wextra` and restore warning suppressions. The legacy `pdnsd` codebase generates numerous warnings that caused the build to fail with strict flags.
- Added `-Wall` and `-Wextra` to `tun2socks` compilation options for Android builds to improve code quality checks, while avoiding `-Werror` to prevent potential breakages in legacy components.
- Kept `tun2socks` strict flags (`-Werror`) for non-Android builds.

+++ native/pdnsd/src/helpers.h
@@ -175,7 +175,7 @@ inline static ssize_t write_all(int fd,const void *data,size_t n)
 {
   ssize_t written=0;
 
-  while(written<n) {
+  while((size_t)written<n) {
       ssize_t m=write(fd,(const void*)(((const char*)data)+written),n-written);
 
       if(m<0)
